working emails with new core mailer
