+++ controller/auth.js
@@ -2,21 +2,17 @@
 
 var passport = require('passport'),
 	path = require('path'),
-	jwt = require('jsonwebtoken'),
-	async = require('async'),
 	fs = require('fs-extra'),
 	merge = require('utils-merge'),
 	Utilities = require('periodicjs.core.utilities'),
 	ControllerHelper = require('periodicjs.core.controller'),
 	Extensions = require('periodicjs.core.extensions'),
-	bcrypt = require('bcrypt'),
 	CoreExtension = new Extensions({
 		extensionFilePath: path.resolve(process.cwd(), './content/config/extensions.json')
 	}),
 	extend = require('util-extend'),
 	CoreUtilities,
 	CoreController,
-	emailtransport,
 	appSettings,
 	mongoose,
 	User,
@@ -99,256 +95,6 @@ var rememberme = function (req, res, next) {
 	}
 	next();
 };
-
-// Utility Functions
-var waterfall = function (array, cb) {
-	async.waterfall(array, cb);
-};
-var encode = function (data) {
-	return jwt.sign(data, loginExtSettings.token.secret);
-};
-
-var decode = function (data, cb) {
-	jwt.verify(data, loginExtSettings.token.secret, {}, function (err, decoded_token) {
-		if (err) {
-			console.log("Error from JWT.verify", err.name);
-			console.log("Error from JWT.verify", err.message);
-		}
-		cb(decoded_token);
-	});
-};
-
-var hasExpired = function (created) {
-	var now = new Date();
-	var diff = (now.getTime() - created);
-	return diff > loginExtSettings.token.ttl;
-};
-
-
-var invalidateUserToken = function (req, res, next, cb) {
-	var token = req.params.token;
-	User.findOne({
-		"attributes.reset_token": token
-	}, function (err, usr) {
-		if (err) {
-			console.log('error finding the user for invalidate token fn');
-			cb(err, null);
-		}
-		usr.attributes.reset_token = "";
-		usr.attributes.reset_token_expires_millis = 0;
-		cb(false, req, res, next, usr);
-	});
-};
-
-var resetPassword = function (req, res, next, user, cb) {
-	var err;
-	if (req.body.password) {
-		if (req.body.password !== req.body.passwordconfirm) {
-			err = new Error('Passwords do not match');
-			req.flash('error', err);
-			cb(err, null);
-		}
-		else if (req.body.password === undefined || req.body.password.length < 8) {
-			err = new Error('Password is too short');
-			req.flash('error', err);
-			cb(err, null);
-		}
-		else {
-			var salt = bcrypt.genSaltSync(10),
-				hash = bcrypt.hashSync(req.body.password, salt);
-			user.password = hash;
-			cb(null, user);
-		}
-	}
-}
-
-/**
- * description The save user function has two special fn calls on the model to mark the properties on it as changed/modified this gets around some werid edge cases when its being updated in memory but not save in mongo
- *
- */
-function saveUser(user, cb) {
-	user.markModified('attributes');
-	user.markModified('password');
-	user.save(function (err, usr) {
-		if (err) {
-			cb(err, null)
-		}
-		cb(null, usr);
-	})
-}
-
-
-var getUser = function (req, res, next, cb) {
-	User.findOne({
-		email: req.body.email
-	}, function (err, user) {
-		if (err) {
-			cb(err, null);
-		}
-		else if (user) {
-			cb(false, user);
-		}
-		else {
-			req.flash('error', 'No user with that email found!')
-			cb(new Error('No user with that email found.'), null);
-		}
-	});
-};
-
-var generateToken = function (user, cb) {
-	//Generate reset token and URL link; also, create expiry for reset token
-	//make sure attributes exists || create it via merge
-	var now = new Date();
-	var expires = new Date(now.getTime() + (loginExtSettings.token.resetTokenExpiresMinutes * 60 * 1000)).getTime();
-	user.attributes = {};
-	user.attributes.reset_token = encode({
-		email: user.email,
-		apikey: user.apikey
-	});
-	user.attributes.reset_token_expires_millis = expires;
-	//TODO: Look into why mongoose properties 
-	//are not being saved during async fn calls
-	user.markModified('attributes');
-	user.save(function (err) {
-		if (err) {
-			cb(err, null);
-		}
-		cb(null, user);
-	});
-};
-
-// create a func for the mail options
-
-var emailConfig = function (user, cb) {
-	var options = {};
-	options.mailtransport = emailtransport;
-	options.subject = "You forgot your password";
-	options.user = user;
-	options.to = user.email;
-	options.replyTo = appSettings.adminnotificationemail;
-	cb(null, options);
-
-}
-
-
-
-var sendEmail = function (options, cb) {
-	//require mailer 
-	var mailtransport = options.mailtransport,
-		user = options.user,
-		mailoptions = {};
-
-	mailoptions.to = (options.to) ? options.to : appSettings.adminnotificationemail
-	mailoptions.cc = options.cc; //options.ccc;
-	mailoptions.bcc = options.bcc;
-	mailoptions.replyTo = options.replyTo;
-	mailoptions.subject = options.subject;
-	if (options.generatetextemail) {
-		mailoptions.generateTextFromHTML = true;
-	}
-	mailoptions.html = options.html;
-	mailoptions.text = options.text;
-	mailtransport.sendMail(mailoptions, cb);
-};
-
-
-//Post to auth/forgot with the users email
-var forgot = function (req, res, next) {
-	var arr = [
-		function (cb) {
-			cb(null, req, res, next)
-		},
-		getUser,
-		generateToken,
-		emailConfig,
-		sendEmail
-	];
-
-	waterfall(arr,
-		function (err, results) {
-			if (err) {
-				req.flash('error', err);
-				res.redirect('/auth/forgot');
-			}
-			res.send(results);
-		});
-};
-
-//GET if the user token is vaild show the change password page
-var reset = function (req, res, next) {
-	var token = req.params.token,
-		current_user,
-		decode_token;
-	var d_token = decode(token, function (decode) {
-		decode_token = decode;
-	});
-
-	//Find the User by their token
-	User.findOne({
-		'attributes.reset_token': token
-	}, function (err, user) {
-		if (err || !user) {
-			req.flash('error', 'Password reset token is invalid.');
-			return res.redirect('/auth/forgot');
-		}
-		current_user = user;
-		//Check to make sure token hasn't expired
-		if (hasExpired(user.attributes.reset_token_expires_millis)) {
-			req.flash('error', 'Password reset token is has expired.');
-			return res.redirect('/auth/forgot');
-		}
-		//Check to make sure token is valid and sign by us
-		if (current_user.email !== decode_token.email && current_user.api_key !== decode_token.api_key) {
-			req.flash('error', 'This token is not valid please try again');
-			res.redirect('/auth/forgot');
-		}
-		CoreController.getPluginViewDefaultTemplate({
-				viewname: 'user/reset',
-				themefileext: appSettings.templatefileextension,
-				extname: 'periodicjs.ext.login'
-			},
-			function (err, templatepath) {
-				CoreController.handleDocumentQueryRender({
-					res: res,
-					req: req,
-					renderView: templatepath,
-					responseData: {
-						pagedata: {
-							title: 'Reset Password',
-							current_user: current_user
-						},
-						user: req.user
-					}
-				});
-			});
-
-	});
-};
-
-
-//POST change the users old password to the new password in the form
-var token = function (req, res, next) {
-	var user_token = req.params.token;
-	waterfall([
-			function (cb) {
-				cb(null, req, res, next);
-			},
-			invalidateUserToken,
-			resetPassword,
-			saveUser,
-			emailConfig,
-			sendEmail,
-		],
-		function (err /*, results*/ ) {
-			if (err) {
-				req.flash('error', 'Opps Something went wrong Please Try Again!');
-				res.redirect('/auth/reset/' + user_token);
-			}
-			req.flash('success', 'Password Sucessfully Changed!');
-			res.redirect('/auth/login');
-		});
-};
-
 
 
 /**
@@ -470,17 +216,10 @@ var controller = function (resources) {
 	passportController.deserialize();
 	passport = passportController.passport;
 	passportController.usePassport();
-	emailController = require('./email_controller')(resources, {
-		User: User
-	});
-
 	return {
 		rememberme: rememberme,
 		login: login,
 		logout: logout,
-		forgot: forgot,
-		reset: reset,
-		token: token,
 		ensureAuthenticated: ensureAuthenticated,
 		loginExtSettings: loginExtSettings,
 		passport: passport
