+++ package.json
@@ -34,7 +34,7 @@
     "passport-twitter": "^1.0.2",
     "periodicjs.core.controller": "~3.1.1",
     "periodicjs.core.extensions": "~4.0.0",
-    "periodicjs.core.mailer": "~3.1.1",
+    "periodicjs.core.mailer": "~4.0.0",
     "periodicjs.core.utilities": "~3.1.1",
     "util-extend": ">=1.0.1"
   },
