+++ views/email/user/forgot.ejs
@@ -1,0 +1,18 @@
+<html>
+<body style="background:#ececec; margin:1em; padding:1em;">
+	<div style="margin:1em; background:#fff; padding:0.5em;">
+		<h2>Forgot Password Notification.</h2>
+		<p>You've chosen to reset your password, click here or copy this url (<%- user.attributes.reset_token %>) to reset your password</p>
+
+		<p> 
+			If you did not request your password to be reset please ignore this email and your password will stay as it is.
+		</p>
+		<p>
+			Login with your email or username (<a href='http://example.com/<%= user.username %>' target='_blank'><%= user.username %></a>) at 
+			<a href='http://example.com/' target='_blank'>
+			http://example.com/
+			</a> 
+		</p>
+	</div>
+</body>
+</html>
