--- views/user/email/forgot.ejs
@@ -1,16 +1,0 @@
-
-<html>
-<body style="background:#ececec; margin:1em; padding:1em;">
-	<div style="margin:1em; background:#fff; padding:0.5em;">
-		<h2>Forgot Password Notification.</h2>
-
-    <p> Hey: <%= user.username %> </p>
-
-    <h2>Reset Email (simulation)</h2>
-    <br><p>To reset your password click the URL below.</p><br>
-    <a href='<%=resetLink %> '> <%= resetLink %></a><br>
-    If you did not request your password to be reset please 
-    ignore this email and your password will stay as it is.
-	</div>
-</body>
-</html>
